+++ BOCCHI.Common/Data/EventDrops/MonsterNote.cs
@@ -1,0 +1,15 @@
+﻿namespace BOCCHI.Common.Data.EventDrops;
+
+public enum MonsterNote : uint
+{
+    CloisterDemon = 47728,
+    CrescentBerserker = 47730,
+    NymianPotaladus = 47731,
+    BlackChocobos = 47732,
+    TradeTortoise = 47733,
+    MythicIdol = 47734,
+    DeadStars = 47735,
+    MarbleDragon = 47736,
+    Magitaur = 47737,
+    PersistentPots = 47738,
+}