+++ BOCCHI/Config/Configuration.cs
@@ -27,6 +27,8 @@ public class Configuration : IConfiguration
 
     public TreasureConfig TreasureConfig { get; set; } = new();
 
+    public EventDropConfig EventDropConfig { get; set; } = new();
+
     public ForkedTowerConfig ForkedTowerConfig { get; set; } = new();
 
     public ShoppingConfig ShoppingConfig { get; set; } = new();